x64 compile fixes (warnings)

+++ branches/psycle-186/psycle/src/Seib-vsthost/CVSTHost.Seib.cpp
@@ -237,11 +237,11 @@ namespace seib {
 		}
 
 		template <class T>
-		bool CFxBase::Read(T &f,bool allowswap)	{ int i=fread(&f,sizeof(T),1,pf); if (NeedsBSwap && allowswap) SwapBytes(f); return (bool)i; }
+		bool CFxBase::Read(T &f,bool allowswap)	{ size_t i=fread(&f,sizeof(T),1,pf); if (NeedsBSwap && allowswap) SwapBytes(f); return (bool)i; }
 		template <class T>
-		bool CFxBase::Write(T f, bool allowswap)	{ if (NeedsBSwap && allowswap) SwapBytes(f); int i=fwrite(&f,sizeof(T),1,pf);  return (bool)i; }
-		bool CFxBase::ReadArray(void* f,int size)	{ int i=fread(f,size,1,pf); return (bool)i; }
-		bool CFxBase::WriteArray(void* f, int size)	{ int i=fwrite(f,size,1,pf);  return (bool)i; }
+		bool CFxBase::Write(T f, bool allowswap)	{ if (NeedsBSwap && allowswap) SwapBytes(f); size_t i=fwrite(&f,sizeof(T),1,pf);  return (bool)i; }
+		bool CFxBase::ReadArray(void* f,int size)	{ size_t i=fread(f,size,1,pf); return (bool)i; }
+		bool CFxBase::WriteArray(void* f, int size)	{ size_t i=fwrite(f,size,1,pf); return (bool)i; }
 		bool CFxBase::ReadHeader()
 		{
 			VstInt32 chunkMagic(0),byteSize(0);
